Note: the branch currently crashes in blender_camera_nodes.cpp:
BL::NodeTree b_ntree = b_data.node_groups[nodes_tree_name];

The crash was introduced in:
cb7cf523e5c000609f32a382e2c0fcc57f635a42

Conflicts:
	intern/cycles/SConscript
	intern/cycles/blender/addon/__init__.py
	intern/cycles/blender/addon/properties.py
	intern/cycles/blender/blender_camera.cpp
	intern/cycles/kernel/kernel_types.h
	intern/cycles/kernel/svm/svm.h
	intern/cycles/kernel/svm/svm_types.h
	intern/cycles/render/camera.cpp
	intern/cycles/render/camera.h

+++ source/blender/imbuf/intern/scaling.c
@@ -33,7 +33,6 @@
 
 
 #include "BLI_utildefines.h"
-#include "BLI_math_base.h"
 #include "BLI_math_color.h"
 #include "BLI_math_interp.h"
 #include "MEM_guardedalloc.h"
@@ -42,7 +41,6 @@
 #include "IMB_imbuf_types.h"
 #include "IMB_imbuf.h"
 
-#include "IMB_allocimbuf.h"
 #include "IMB_filter.h"
 
 #include "BLI_sys_types.h" // for intptr_t support
@@ -1557,7 +1555,7 @@ struct ImBuf *IMB_scalefastImBuf(struct ImBuf *ibuf, unsigned int newx, unsigned
 	struct imbufRGBA *rectf, *_newrectf, *newrectf;
 	int x, y;
 	bool do_float = false, do_rect = false;
-	int ofsx, ofsy, stepx, stepy;
+	size_t ofsx, ofsy, stepx, stepy;
 
 	rect = NULL; _newrect = NULL; newrect = NULL;
 	rectf = NULL; _newrectf = NULL; newrectf = NULL;
